+++ tinymall-db/src/main/java/com/attitude/tinymall/service/impl/LitemallDeliveryDetailServiceImpl.java
@@ -100,7 +100,7 @@ public class LitemallDeliveryDetailServiceImpl implements LitemallDeliveryDetail
 
         ResponseEntity<AddOrderResult> res = remoteDadaDeliveryClient.addOrder(orderParams);
         if (res.isSuccess()) {
-            order.setDeliverFee(res.getResult().getFee().intValue());
+            order.setDeliverFee(res.getResult().getFee());
             LitemallDeliveryDetail litemallDeliveryDetail = new LitemallDeliveryDetail();
             litemallDeliveryDetail.setDeliveryId(deliveryId);
             litemallDeliveryDetail.setDistance("" + res.getResult().getDistance());
